+++ Code/Libraries/Foundation/Serialization/Text.hpp
@@ -18,12 +18,12 @@ public:
             DataVersion::Enum version = DataVersion::Current,
             FileMode::Enum fileMode = FileMode::Write);
   bool OpenBuffer(DataVersion::Enum version = DataVersion::Current, FileMode::Enum fileMode = FileMode::Write);
-  uint GetBufferSize();
+  size_t GetBufferSize();
   String GetString();
 
   SerializerClass::Enum GetClass() override;
 
-  void ExtractInto(::byte* data, uint size);
+  void ExtractInto(::byte* data, size_t size);
   DataBlock ExtractAsDataBlock();
   void Close();
 
@@ -52,10 +52,14 @@ public:
   bool StringField(cstr typeName, cstr fieldName, StringRange& stringRange) override;
 
   // Array Serialization
-  bool ArrayField(
-      cstr typeName, cstr fieldName, ::byte* data, ArrayType arrayType, uint numberOfElements, uint sizeOftype) override;
+  bool ArrayField(cstr typeName,
+                  cstr fieldName,
+                  ::byte* data,
+                  ArrayType arrayType,
+                  size_t numberOfElements,
+                  size_t sizeOftype) override;
 
-  void ArraySize(uint& arraySize){};
+  void ArraySize(uint& arraySize) {};
 
   // Enum Serialization
   bool EnumField(cstr enumTypeName, cstr fieldName, uint& enumValue, BoundType* type) override;
@@ -67,7 +71,7 @@ public:
   // private:
   StringBuilder mStream;
   String mFilename;
-  uint mDepth;
+  size_t mDepth;
   FileMode::Enum mWriteMode;
 
   /// The version being saved
